添加注册激活功能

+++ yang-auth/src/main/java/cn/codeyang/auth/mapper/AuthorityMapper.java
@@ -1,7 +1,6 @@
 package cn.codeyang.auth.mapper;
 
 import cn.codeyang.auth.entity.Authority;
-import cn.codeyang.auth.entity.User;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.springframework.stereotype.Repository;
 
@@ -16,5 +15,5 @@ public interface AuthorityMapper extends BaseMapper<Authority> {
 	 * 根据userId查询用户权限
 	 * @param userId 用户ID
 	 */
-	List<Authority> findAuthoritiesByUserId(Integer userId);
+	List<Authority> findAuthoritiesByUserId(Long userId);
 }
