+++ yang-auth/src/main/java/cn/codeyang/auth/service/OrganizationService.java
@@ -11,7 +11,7 @@ public interface OrganizationService extends IService<Organization> {
 
 	boolean createOrganization(Organization organization);
 
-	void deleteOrganization(Integer id);
+	void deleteOrganization(Long id);
 
 	Organization selectOne(Organization organization);
 }
