+++ yang-auth/src/main/java/cn/codeyang/auth/service/impl/AuthorityServiceImpl.java
@@ -20,7 +20,7 @@ public class AuthorityServiceImpl extends ServiceImpl<AuthorityMapper, Authority
 	 * 根据userId查询用户权限
 	 * @param userId 用户ID
 	 */
-	public List<Authority> findAuthoritiesByUserId(Integer userId){
+	public List<Authority> findAuthoritiesByUserId(Long userId){
 		return authorityMapper.findAuthoritiesByUserId(userId);
 	}
 }
